Implement registration system with associated database migrations, models, controllers, and routes for managing classes, blocks, fields, and plans.

+++ db/migration/v1.0/20251115002641_create_blocks_table.js
@@ -4,26 +4,30 @@ import blocksModule from '../../seeds/blocksData.js';
 const { getBlocksData } = blocksModule;
 
 export async function up(knex) {
-    await knex.schema.createTable('blocks', (table) => {
-        table.increments('id').primary();
-        table
-            .integer('module_id')
-            .unsigned()
-            .references('id')
-            .inTable('modules')
-            .onDelete('CASCADE');
-        table.string('name').notNullable();
-        table.string('description');
-        table.integer('order').defaultTo(0);
-        table.boolean('collapsible').defaultTo(false);
-        table.string('display_mode').defaultTo('edit');
-        table.timestamps(true, true);
-    });
+    const exists = await knex.schema.hasTable('blocks');
 
-    console.log("Table 'blocks' created successfully.");
+    if (!exists) {
+        await knex.schema.createTable('blocks', (table) => {
+            table.increments('id').primary();
+            table
+                .integer('module_id')
+                .unsigned()
+                .references('id')
+                .inTable('modules')
+                .onDelete('CASCADE');
+            table.string('name').notNullable();
+            table.string('description');
+            table.integer('order').defaultTo(0);
+            table.boolean('collapsible').defaultTo(false);
+            table.string('display_mode').defaultTo('edit');
+            table.timestamps(true, true);
+        });
 
-    const blocksData = await getBlocksData(knex);
-    await knex('blocks').insert(blocksData);
+        console.log("Table 'blocks' created successfully.");
+
+        const blocksData = await getBlocksData(knex);
+        await knex('blocks').insert(blocksData);
+    }
 };
 
 export async function down(knex) {
